rename raftstore
