+++ components/raftstore-v2/tests/integrations/test_status.rs
@@ -5,7 +5,7 @@ use std::assert_matches::assert_matches;
 use futures::executor::block_on;
 use kvproto::raft_cmdpb::{RaftCmdRequest, StatusCmdType};
 use raftstore_v2::router::{PeerMsg, PeerTick, QueryResChannel, QueryResult, RaftRequest};
-use tikv_util::raftstore::new_peer;
+use tikv_util::store::new_peer;
 
 use crate::Cluster;
