+++ components/raftstore/src/store/fsm/store.rs
@@ -51,8 +51,9 @@ use tikv_util::{
     future::poll_future_notify,
     info, is_zero_duration,
     mpsc::{self, LooseBoundedSender, Receiver},
-    raftstore::find_peer,
-    slow_log, sys as sys_util,
+    slow_log,
+    store::find_peer,
+    sys as sys_util,
     sys::disk::{get_disk_status, DiskUsage},
     time::{duration_to_sec, Instant as TiInstant},
     timer::SteadyTimer,
