+++ components/tikv_util/src/lib.rs
@@ -53,7 +53,7 @@ pub mod memory;
 pub mod metrics;
 pub mod mpsc;
 pub mod quota_limiter;
-pub mod raftstore;
+pub mod store;
 pub mod stream;
 pub mod sys;
 pub mod thread_group;
