+++ tests/integrations/raftstore/test_conf_change.rs
@@ -21,7 +21,7 @@ use raft::eraftpb::{ConfChangeType, MessageType};
 use raftstore::Result;
 use test_pd_client::TestPdClient;
 use test_raftstore::*;
-use tikv_util::{config::ReadableDuration, raftstore::is_learner, time::Instant, HandyRwLock};
+use tikv_util::{config::ReadableDuration, store::is_learner, time::Instant, HandyRwLock};
 
 fn test_simple_conf_change<T: Simulator>(cluster: &mut Cluster<T>) {
     let pd_client = Arc::clone(&cluster.pd_client);
