+++ tests/integrations/raftstore/test_region_change_observer.rs
@@ -17,7 +17,7 @@ use raftstore::coprocessor::{
 };
 use test_raftstore::{new_node_cluster, Cluster, NodeCluster};
 use tikv_util::{
-    raftstore::{find_peer, new_peer},
+    store::{find_peer, new_peer},
     HandyRwLock,
 };
